Move guest auth to root path
